wip: adding frontend ai provider
