+++ backend/windmill-api/src/workspaces.rs
@@ -695,7 +695,6 @@ async fn edit_copilot_config(
 
 #[derive(Serialize)]
 struct CopilotInfo {
-    pub ai_resource_type: String,
     pub exists_ai_resource: bool,
     pub code_completion_enabled: bool,
 }
@@ -713,19 +712,13 @@ async fn get_copilot_info(
     .map_err(|e| Error::InternalErr(format!("getting ai_resource and code_completion_enabled: {e:#}")))?;
     tx.commit().await?;
 
-    let ai_resource_type = if let Some(ai_resource) = record.ai_resource {
-        if let Ok(ai_resource) = serde_json::from_value::<AiRessource>(ai_resource) {
-            ai_resource.provider
-        } else {
-            "".to_string()
-        }
+    let exists_ai_resource = if let Some(ai_resource) = record.ai_resource {
+        serde_json::from_value::<AiRessource>(ai_resource).is_ok()
     } else {
-        "".to_string()
+        false
     };
 
-    let exists_ai_resource = !ai_resource_type.is_empty();
     Ok(Json(CopilotInfo {
-        ai_resource_type,
         exists_ai_resource,
         code_completion_enabled: record.code_completion_enabled,
     }))
