+++ frontend/src/lib/components/Dev.svelte
@@ -113,8 +113,7 @@
 				copilotInfo.set(await WorkspaceService.getCopilotInfo({ workspace }))
 			} catch (err) {
 				copilotInfo.set({
-					ai_resource_type: '',
-					exists_ai_resource_path: false,
+					exists_ai_resource: false,
 					code_completion_enabled: false
 				})
