+++ frontend/src/lib/components/copilot/FlowCopilotStatus.svelte
@@ -31,7 +31,7 @@
 	$: copilotStatus && copilotPopover?.refresh()
 </script>
 
-{#if $copilotInfo.exists_ai_resource_path}
+{#if $copilotInfo.exists_ai_resource}
 	<ManualPopover bind:this={copilotPopover}>
 		<Button
 			size="xs"
@@ -137,7 +137,7 @@
 		<div class="block text-primary">
 			<p class="text-sm"
 				>Enable Windmill AI in the <a
-					href="{base}/workspace_settings?tab=openai"
+					href="{base}/workspace_settings?tab=ai"
 					target="_blank"
 					class="inline-flex flex-row items-center gap-1"
 					>workspace settings <ExternalLink size={16} /></a
