+++ frontend/src/lib/components/flows/content/FlowInputsQuick.svelte
@@ -86,7 +86,7 @@
 	}
 
 	async function onGenerate() {
-		if (!selectedCompletion && !$copilotInfo.exists_ai_resource_path) {
+		if (!selectedCompletion && !$copilotInfo.exists_ai_resource) {
 			sendUserToast(
 				'Windmill AI is not enabled, you can activate it in the workspace settings',
 				true
