+++ frontend/src/routes/(root)/(logged)/+layout.svelte
@@ -234,7 +234,7 @@
 			console.log(err)
 			copilotInfo.set({
 				ai_resource_type: '',
-				exists_ai_resource_path: false,
+				exists_ai_resource: false,
 				code_completion_enabled: false
 			})
 			console.error('Could not get copilot info')
